src/2024/15/2/level2.test.ts
import { describe, expect, test } from "vitest";
import { solve } from "./level2";
import { useInput, useProdInput } from "../../../utils/test-utils";

describe("level2", () => {
  test("../example.txt", async () => {
    const input = await useInput(__dirname);
    const output = solve(input);
    expect(output).toBe(9021);
  });

  test("../example2.txt", async () => {
    const input = await useInput(__dirname);
    const output = solve(input);
    expect(output).toBe(618);
  });

  test("../prod.txt", async () => {
    const input = await useProdInput(__dirname);
    const output = solve(input);
    expect(output).toBe(1519202);
  });
});
